+++ ChangeLog.md
@@ -5,6 +5,7 @@ AWS Core change log
 
 ## 3.2.0 / 2026-04-23
 
+* Added `com.amazon.aws.Credentials::environment()` - @thekid
 * Made compatible with `xp-framework/http` version 11.0 - @thekid
 
 ## 3.1.0 / 2026-04-10